# docs/api/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "The Swagger API documentation of the WordsApp",
        "version": "0.0.1"
    },
    "servers": [
        {
      "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BasicAuth": []
        }
    ],
...
    "paths": {
...
        "/api/accounts": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get an account",
                "operationId": "get_1",
                "responses": {
                    "200": {
                        "description": "Account found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Update existing account",
                "operationId": "update_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAccountRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Create a new account and return Base64-encoded auth token",
                "operationId": "public",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveAccountDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Account created and token returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthBaseTokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Account already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthBaseTokenResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete account",
                "description": "Delete an account even if it doesn't exist",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
...
        "/api/accounts/login64": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Login and return Base64-encoded auth token",
                "operationId": "public_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Token returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthBaseTokenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthBaseTokenResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateAccountRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "username"
                ]
            },
            "AccountDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "id",
                    "username"
                ]
            },
...
            "SaveAccountDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password",
                    "username"
                ]
            },
            "AuthBaseTokenResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "AccountLoginRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
...
        },
        "securitySchemes": {
            "BasicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}